Fixes to subscription tracking
